updated controll and views vehicles

+++ app/Http/Controllers/VeiculosController.php
@@ -17,14 +17,25 @@ class VeiculosController extends Controller
     {       
         return view('veiculos.create');
     }
+
+    public function verificaUser($veiculo){
+        $user = Veiculos::with('usuario')->find($veiculo);
+        if ($user->usuario->id != Auth::id()){
+            return true;
+        }
+    }
+
     public function store(Request $request)
-    {   
+    {  
+        $mensagens = [
+            'formato_placa_de_veiculo' => 'A placa deve conter o seguinte formato: ABC1D23'
+        ];
         $this->validate($request, [
-            'placa' => 'required',
+            'placa' => 'required|formato_placa_de_veiculo|unique:veiculos',
             'modelo' => 'required',
             'cor' => 'required',  
             'tipo' => 'required',               
-        ]);      
+        ],$mensagens);      
 
         try {
             $input = $request->all();
@@ -44,21 +55,32 @@ class VeiculosController extends Controller
     }
     public function edit($id)
     {   
+        if ($this->verificaUser($id) == true){
+            return redirect()->route('veiculos.index')
+            ->withErrors('Veiculos não pertence ao usuario autenticado');
+        };
         $veiculos = Veiculos::find($id);
-        return view('veiculos.create', compact('veiculos'));
+        return view('veiculos.edit', compact('veiculos'));
     }
-    public function update(Request $request, $id){    
+    public function update(Request $request, $id){   
+        if ($this->verificaUser($id) == true){
+            return redirect()->route('veiculos.index')
+            ->withErrors('Veiculos não pertence ao usuario autenticado');
+        };
         $veiculos = Veiculos::find($id);   
+        $mensagens = [
+            'formato_placa_de_veiculo' => 'A placa deve conter o seguinte formato: ABC1D23'
+        ];
         $this->validate($request, [
-            'placa' => 'required|unique:placa',
+            'placa' => 'required|formato_placa_de_veiculo|unique:veiculos,placa,'. $id,
             'modelo' => 'required',
             'cor' => 'required',  
             'tipo' => 'required',               
-        ]);   
+        ], $mensagens);   
                 
-        try {
+        try {            
             $input = $request->all();
-            $input['user_id'] = Auth::id();
+            $input['user_id'] = Auth::id();           
             $veiculos->update($input); 
             return redirect()->route('veiculos.index')
             ->with('success','Veiculos atualizado com sucesso');
@@ -69,7 +91,11 @@ class VeiculosController extends Controller
     }
     public function destroy($id)
     {       
-        try {
+        if ($this->verificaUser($id) == true){
+            return redirect()->route('veiculos.index')
+            ->withErrors('Veiculos não pertence ao usuario autenticado');
+        };
+        try {            
             $veiculos = Veiculos::find($id);
             $veiculos->delete(); 
             return redirect()->route('veiculos.index')
